Added puzzle 4
Added a puzzle wherein the correct color pool balls must be placed on the table at color coordinated spots when the ball is placed in the correct spot the ball locks in place and when all three cards are in place a card falls from the sky onto the center of the table. Still need to add more pool balls as red herrings.